improving error message for timeout
